Main Activity and Main Fragment: change FragmentContainer to disassociate activity and main
